requirements.txt
# Core dependencies
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.17.0

# Database
sqlalchemy>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# API requests
requests>=2.31.0

# File processing
openpyxl>=3.1.0  # For Excel files
xlrd>=2.0.0      # For older Excel formats

# Optional but recommended
altair>=5.0.0    # Alternative charts
numpy>=1.24.0    # Numerical operations
Pillow>=10.0.0   # Image processing
